refactor: extract admin, superadmin, health, metrics into per-app packages
Fifth phase — the role-scoped admin consoles, mechanical moves using the
now-proven recipe:

- app/blueprints/admin.py + its 4 flat admin_*.html templates ->
  app/admin/ (templates namespaced into templates/admin/, matching the
  render_template('admin/...') paths)
- app/blueprints/superadmin.py -> app/superadmin/, and the whole already-
  namespaced app/templates/superadmin/ folder moves with it
- app/blueprints/health.py -> app/health/, app/blueprints/metrics.py ->
  app/metrics/ — both render 'superadmin/*.html' templates without
  owning a template_folder themselves; verified this still resolves
  correctly, since Flask merges every registered blueprint's
  template_folder into one shared Jinja search path (confirmed via
  test_request_context: rendering 'superadmin/health.html' finds the
  file physically living under app/superadmin/templates/, not health/)

app/blueprints/ now contains only courses.py, main.py, and discussion.py
— the three files slated for the courses.py/main.py decomposition in
the remaining phases.

Verified: create_app() builds with all 224 routes intact, /healthz
returns 200, and pytest is unchanged from baseline (25 pre-existing
failures, 99 passing).
