+++ app/admin/templates/admin/admin_class_detail.html
