+++ app/admin/templates/admin/admin_classes.html
