+++ app/superadmin/routes.py
@@ -13,7 +13,7 @@ from app.services.email_service import send_school_approved_email, send_ticket_u
 from app.tickets.services import transition_status, TicketStatus as TStat
 from app.core.middleware import invalidate_school_cache
 
-superadmin_bp = Blueprint('superadmin', __name__, url_prefix='/superadmin')
+superadmin_bp = Blueprint('superadmin', __name__, url_prefix='/superadmin', template_folder='templates')
 
 
 def validate_password(password):
